# docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/foods/search": {
      "get": {
        "operationId": "FoodController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "food"
        ]
      }
    },
    "/foods/barcode/{upc}": {
      "get": {
        "operationId": "FoodController_getByBarcode",
        "parameters": [
          {
            "name": "upc",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "food"
        ]
      }
    },
    "/foods/{id}": {
      "get": {
        "operationId": "FoodController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "food"
        ]
      }
    },
    "/foods": {
      "post": {
        "operationId": "FoodController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFoodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "food"
        ]
      }
    },
...
    "/community/leaderboard": {
      "get": {
        "operationId": "CommunityController_getLeaderboard",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ]
      }
    },
...
    "/community/wins": {
      "post": {
        "operationId": "CommunityController_postWin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostWinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ]
      }
    },
    "/community/wins/{id}/react": {
      "post": {
        "operationId": "CommunityController_reactToWin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactToWinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ]
      }
    },
...
    "/lists/{type}": {
      "get": {
        "operationId": "ListsController_getList",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lists"
        ]
      },
      "post": {
        "operationId": "ListsController_addItem",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddListItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "lists"
        ]
      }
    },
    "/lists/{type}/clear-checked": {
      "post": {
        "operationId": "ListsController_clearChecked",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lists"
        ]
      }
    },
    "/lists/{type}/bulk": {
      "post": {
        "operationId": "ListsController_bulkAdd",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "lists"
        ]
      }
    },
    "/lists/items/{id}": {
      "patch": {
        "operationId": "ListsController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lists"
        ]
      },
      "delete": {
        "operationId": "ListsController_deleteItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "lists"
        ]
      }
    },
...
    "/v1/admin/coaches/{id}/start-subscription": {
      "post": {
        "operationId": "OwnerBillingController_startSubscription",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "billing-admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Growth Project API",
    "description": "HTTP API for The Growth Project — coach/client SaaS platform covering auth, billing, messaging, check-ins, meal plans, workouts, habits, and admin/federation surfaces. Consumed by the React Native mobile app, the coach console BFF, and partner integrations.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.trygrowthproject.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase-issued JWT. Send as `Authorization: Bearer <token>`. Tokens are verified against the Supabase JWKS in JwtAuthGuard."
      }
    },
    "schemas": {
...
      "CreateFoodDto": {
        "type": "object",
        "properties": {}
      },
...
      "PostWinDto": {
        "type": "object",
        "properties": {}
      },
      "ReactToWinDto": {
        "type": "object",
        "properties": {}
      },
...
      "AddListItemDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateListItemDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}